Continue work on MemoryTracker
There is a mysterious issue where the project will fail to compile with a "TypeInfo cannot be used with -betterC" message originating in "std\array.d(3266)".

Run the compiler a couple times, and the error will disappear and things will compile just fine. I have no idea why this is happening.